Changed API slightly
AddSub now takes a Configer rather than a *Config. The *Config version
is now named AddSubConf. AddSub was added because it should not be
required to keep track of *Config's in user code.

AddSub and AddSubConf both now return an error instead of a *Config, equal
named sub-configurations are now disallowed and will return an error from
both of these functions.