Fixed bug with the new feature.
The new feature to unmarshal after-load assumed that the JSON actually
had more than one item in it. This caused nothing to happen when
len(c.Raws) == 0; Fixed by using a Loaded boolean field to keep
track of pre/post-load state.

This also adds the missing Default calls when no JSON is found.